+++ src/main/java/ru/alfabank/steps/DefaultSteps1.java
@@ -1,0 +1,104 @@
+/**
+ * Copyright 2017 Alfa Laboratory
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package ru.alfabank.steps;
+
+import com.codeborne.selenide.Condition;
+import com.codeborne.selenide.Selenide;
+import com.codeborne.selenide.SelenideElement;
+import com.codeborne.selenide.WebDriverRunner;
+import cucumber.api.java.ru.*;
+import lombok.extern.slf4j.Slf4j;
+import org.apache.commons.lang3.RandomStringUtils;
+import org.hamcrest.Matchers;
+import org.junit.Assert;
+import org.openqa.selenium.By;
+import org.openqa.selenium.Dimension;
+import org.openqa.selenium.Keys;
+import org.openqa.selenium.interactions.Actions;
+import ru.alfabank.alfatest.cucumber.api.AkitaScenario;
+
+import java.awt.*;
+import java.awt.datatransfer.Clipboard;
+import java.awt.datatransfer.ClipboardOwner;
+import java.awt.datatransfer.StringSelection;
+import java.io.File;
+import java.text.SimpleDateFormat;
+import java.util.ArrayList;
+import java.util.HashSet;
+import java.util.List;
+import java.util.Random;
+import java.util.regex.Matcher;
+import java.util.regex.Pattern;
+import java.util.stream.Collectors;
+
+import static com.codeborne.selenide.Condition.not;
+import static com.codeborne.selenide.Selenide.*;
+import static com.codeborne.selenide.WebDriverRunner.getWebDriver;
+import static com.codeborne.selenide.WebDriverRunner.url;
+import static java.lang.String.format;
+import static java.util.Objects.isNull;
+import static java.util.stream.Collectors.toList;
+import static org.hamcrest.MatcherAssert.assertThat;
+import static org.hamcrest.Matchers.*;
+import static org.junit.Assert.*;
+import static ru.alfabank.alfatest.cucumber.ScopedVariables.resolveVars;
+import static ru.alfabank.tests.core.helpers.PropertyLoader.*;
+
+/**
+ * В akitaScenario используется хранилище переменных. Для сохранения/изъятия переменных используются методы setVar/getVar
+ * Каждая страница, с которой предполагается взаимодействие, должна быть описана в соответствующем классе,
+ * наследующем AkitaPage. Для каждого элемента следует задать имя на русском, через аннотацию @Name, чтобы искать
+ * можно было именно по русскому описанию, а не по селектору. Селекторы следует хранить только в классе страницы,
+ * не в степах, в степах - взаимодействие по русскому названию элемента.
+ */
+@Slf4j
+public class DefaultSteps1 {
+
+    private AkitaScenario akitaScenario = AkitaScenario.getInstance();
+
+    private static final int DEFAULT_TIMEOUT = loadPropertyInt("waitingCustomElementsTimeout", 10000);
+
+
+
+//    /**
+//     * Возвращает значение из property файла, если отсутствует, то из пользовательских переменных,
+//     * если и оно отсутствует, то возвращает значение переданной на вход переменной
+//     *
+//     * @return
+//     */
+//    public String getPropertyOrStringVariableOrValue(String propertyNameOrVariableNameOrValue) {
+//        String propertyValue = tryLoadProperty(propertyNameOrVariableNameOrValue);
+//        String variableValue = (String) akitaScenario.tryGetVar(propertyNameOrVariableNameOrValue);
+//
+//        boolean propertyCheck = checkResult(propertyValue, "Переменная " + propertyNameOrVariableNameOrValue + " из property файла");
+//        boolean variableCheck = checkResult(variableValue, "Переменная сценария " + propertyNameOrVariableNameOrValue);
+//
+//        return propertyCheck ? propertyValue : (variableCheck ? variableValue : propertyNameOrVariableNameOrValue);
+//    }
+//
+//    private boolean checkResult(String result, String message) {
+//        if (isNull(result)) {
+//            log.warn(message + " не найдена");
+//            return false;
+//        }
+//        log.info(message + " = " +  result);
+//        akitaScenario.write(message + " = " + result);
+//        return true;
+//    }
+
+
+
+}
