+++ src/main/java/ru/alfabank/steps/defaultSteps/WebPageSteps.java
@@ -1,0 +1,467 @@
+package ru.alfabank.steps.defaultSteps;
+
+import com.codeborne.selenide.Selenide;
+import com.codeborne.selenide.SelenideElement;
+import com.codeborne.selenide.WebDriverRunner;
+import cucumber.api.java.ru.*;
+import lombok.extern.slf4j.Slf4j;
+import org.hamcrest.Matchers;
+import org.openqa.selenium.By;
+import org.openqa.selenium.Dimension;
+import org.openqa.selenium.Keys;
+import org.openqa.selenium.interactions.Actions;
+import ru.alfabank.alfatest.cucumber.api.AkitaScenario;
+
+import java.io.File;
+import java.util.List;
+import java.util.stream.Collectors;
+
+import static com.codeborne.selenide.Selenide.*;
+import static com.codeborne.selenide.WebDriverRunner.getWebDriver;
+import static com.codeborne.selenide.WebDriverRunner.url;
+import static java.util.Objects.isNull;
+import static org.hamcrest.MatcherAssert.assertThat;
+import static org.hamcrest.Matchers.equalTo;
+import static org.hamcrest.Matchers.equalToIgnoringCase;
+import static org.hamcrest.Matchers.is;
+import static org.junit.Assert.assertFalse;
+import static org.junit.Assert.assertNotNull;
+import static org.junit.Assert.assertTrue;
+import static ru.alfabank.alfatest.cucumber.ScopedVariables.resolveVars;
+import static ru.alfabank.tests.core.helpers.PropertyLoader.*;
+
+
+/**
+ * Шаги для работы с вэб-страницей, браузером, переменными и property-файлами, доступные по умолчанию в каждом новом проекте.
+ *
+ * В akitaScenario используется хранилище переменных. Для сохранения/изъятия переменных используются методы setVar/getVar
+ * Каждая страница, с которой предполагается взаимодействие, должна быть описана в соответствующем классе,
+ * наследующем AkitaPage. Для каждого элемента следует задать имя на русском, через аннотацию @Name, чтобы искать
+ * можно было именно по русскому описанию, а не по селектору. Селекторы следует хранить только в классе страницы,
+ * не в степах, в степах - взаимодействие по русскому названию элемента.
+ */
+
+@Slf4j
+public class WebPageSteps {
+
+    private AkitaScenario akitaScenario = AkitaScenario.getInstance();
+
+
+    /**
+     * Проверка того, что все элементы, которые описаны в классе страницы с аннотацией @Name,
+     * но без аннотации @Optional появились на странице
+     * в течение WAITING_APPEAR_TIMEOUT, которое равно значению свойства "waitingAppearTimeout"
+     * из application.properties. Если свойство не найдено, время таймаута равно 8 секундам
+     */
+    @Тогда("^(?:страница|блок|форма|вкладка) \"([^\"]*)\" (?:загрузилась|загрузился)$")
+    public void loadPage(String nameOfPage) {
+        akitaScenario.setCurrentPage(akitaScenario.getPage(nameOfPage));
+        akitaScenario.getCurrentPage().appeared();
+    }
+
+    /**
+     * Проверка того, что все элементы, которые описаны в классе страницы с аннотацией @Name,
+     * но без аннотации @Optional, не появились на странице
+     */
+    @Тогда("^(?:страница|блок|форма|вкладка) \"([^\"]*)\" не (?:загрузилась|загрузился)$")
+    public void loadPageFailed(String nameOfPage) {
+        akitaScenario.setCurrentPage(akitaScenario.getPage(nameOfPage));
+        akitaScenario.getCurrentPage().disappeared();
+    }
+
+    /**
+     * Значение заданной переменной из application.properties сохраняется в переменную в akitaScenario
+     * для дальнейшего использования
+     */
+    @И("^сохранено значение \"([^\"]*)\" из property файла в переменную \"([^\"]*)\"$")
+    public void saveValueToVar(String propertyVariableName, String variableName) {
+        propertyVariableName = loadProperty(propertyVariableName);
+        akitaScenario.setVar(variableName, propertyVariableName);
+        akitaScenario.write("Значение сохраненной переменной " + propertyVariableName);
+    }
+
+    /**
+     * Устанавливается значение переменной в хранилище переменных. Один из кейсов: установка login пользователя
+     */
+    @И("^установлено значение переменной \"([^\"]*)\" равным \"(.*)\"$")
+    public void setVariable(String variableName, String value) {
+        value = getPropertyOrValue(value);
+        akitaScenario.setVar(variableName, value);
+    }
+
+    /**
+     * Проверка равенства двух переменных из хранилища
+     */
+    @Тогда("^значения в переменных \"([^\"]*)\" и \"([^\"]*)\" совпадают$")
+    public void compareTwoVariables(String firstVariableName, String secondVariableName) {
+        String firstValueToCompare = akitaScenario.getVar(firstVariableName).toString();
+        String secondValueToCompare = akitaScenario.getVar(secondVariableName).toString();
+        assertThat(String.format("Значения в переменных [%s] и [%s] не совпадают", firstVariableName, secondVariableName),
+            firstValueToCompare, equalTo(secondValueToCompare));
+    }
+
+    /**
+     * Проверка неравенства двух переменных из хранилища
+     */
+    @Тогда("^значения в переменных \"([^\"]*)\" и \"([^\"]*)\" не совпадают$")
+    public void checkingTwoVariablesAreNotEquals(String firstVariableName, String secondVariableName) {
+        String firstValueToCompare = akitaScenario.getVar(firstVariableName).toString();
+        String secondValueToCompare = akitaScenario.getVar(secondVariableName).toString();
+        assertThat(String.format("Значения в переменных [%s] и [%s] совпадают", firstVariableName, secondVariableName),
+            firstValueToCompare, Matchers.not(equalTo(secondValueToCompare)));
+    }
+
+    /**
+     * Выполняется обновление страницы
+     */
+    @И("^выполнено обновление текущей страницы$")
+    public void refreshPage() {
+        refresh();
+    }
+
+    /**
+     * Выполняется переход по заданной ссылке,
+     * ссылка берется из property / переменной, если такая переменная не найдена,
+     * то берется переданное значение
+     * при этом все ключи переменных в фигурных скобках
+     * меняются на их значения из хранилища akitaScenario
+     */
+    @Когда("^совершен переход по ссылке \"([^\"]*)\"$")
+    public void goToUrl(String address) {
+        String url = resolveVars(getPropertyOrStringVariableOrValue(address));
+        open(url);
+        akitaScenario.write("Url = " + url);
+    }
+
+    /**
+     * Проверка, что текущий URL совпадает с ожидаемым
+     * (берется из property / переменной, если такая переменная не найдена,
+     * то берется переданное значение)
+     */
+    @Тогда("^текущий URL равен \"([^\"]*)\"$")
+    public void checkCurrentURL(String url) {
+        String currentUrl = url();
+        String expectedUrl = resolveVars(getPropertyOrStringVariableOrValue(url));
+        assertThat("Текущий URL не совпадает с ожидаемым", currentUrl, is(expectedUrl));
+    }
+
+    /**
+     * Проверка, что текущий URL не совпадает с ожидаемым
+     * (берется из property / переменной, если такая переменная не найдена,
+     * то берется переданное значение)
+     */
+    @Тогда("^текущий URL не равен \"([^\"]*)\"$")
+    public void checkCurrentURLIsNotEquals(String url) {
+        String currentUrl = url();
+        String expectedUrl = resolveVars(getPropertyOrStringVariableOrValue(url));
+        assertThat("Текущий URL совпадает с ожидаемым", currentUrl, Matchers.not(expectedUrl));
+    }
+
+    /**
+     * Выполняется переход по заданной ссылке.
+     * Шаг содержит проверку, что после перехода загружена заданная страница.
+     * Ссылка может передаваться как строка, так и как ключ из application.properties
+     * Deprecated
+     */
+    @Deprecated
+    @И("^совершен переход на страницу \"([^\"]*)\" по ссылке из property файла \"([^\"]*)\"$")
+    public void goToSelectedPageByLinkFromPropertyFile(String pageName, String urlOrName) {
+        String address = loadProperty(urlOrName, resolveVars(urlOrName));
+        akitaScenario.write(" url = " + address);
+        open(address);
+        loadPage(pageName);
+    }
+
+    /**
+     * Выполняется переход по заданной ссылке.
+     * Шаг содержит проверку, что после перехода загружена заданная страница.
+     * Ссылка может передаваться как строка, так и как ключ из application.properties
+     */
+    @И("^совершен переход на страницу \"([^\"]*)\" по ссылке \"([^\"]*)\"$")
+    public void goToSelectedPageByLink(String pageName, String urlOrName) {
+        String address = loadProperty(urlOrName, resolveVars(urlOrName));
+        akitaScenario.write(" url = " + address);
+        open(address);
+        loadPage(pageName);
+    }
+
+    /**
+     * Ожидание в течение заданного количества секунд
+     */
+    @Когда("^выполнено ожидание в течение (\\d+) (?:секунд|секунды)")
+    public void waitForSeconds(long seconds) {
+        sleep(1000 * seconds);
+    }
+
+    /**
+     * Проверка того, что блок исчез/стал невидимым
+     */
+    @Тогда("^(?:страница|блок|форма) \"([^\"]*)\" (?:скрыт|скрыта)")
+    public void blockDisappeared(String nameOfPage) {
+        akitaScenario.getPage(nameOfPage).disappeared();
+    }
+
+    /**
+     * Эмулирует нажатие клавиш на клавиатуре
+     */
+    @И("^выполнено нажатие на клавиатуре \"([^\"]*)\"$")
+    public void pushButtonOnKeyboard(String buttonName) {
+        Keys key = Keys.valueOf(buttonName.toUpperCase());
+        switchTo().activeElement().sendKeys(key);
+    }
+
+    /**
+     * Эмулирует нажатие сочетания клавиш на клавиатуре.
+     * Допустим, чтобы эмулировать нажатие на Ctrl+A, в таблице должны быть следующие значения
+     * | CONTROL |
+     * | a       |
+     *
+     * @param keyNames название клавиши
+     */
+    @И("^выполнено нажатие на сочетание клавиш из таблицы$")
+    public void pressKeyCombination(List<String> keyNames) {
+        Iterable<CharSequence> listKeys = keyNames.stream()
+            .map(this::getKeyOrCharacter)
+            .collect(Collectors.toList());
+        String combination = Keys.chord(listKeys);
+        switchTo().activeElement().sendKeys(combination);
+    }
+
+    private CharSequence getKeyOrCharacter(String key) {
+        try {
+            return Keys.valueOf(key.toUpperCase());
+        } catch (IllegalArgumentException ex) {
+            return key;
+        }
+    }
+
+    /**
+     * Устанавливает размеры окна браузера
+     * Deprecated
+     */
+    @Deprecated
+    @И("^установить разрешение экрана \"([^\"]*)\" ширина и \"([^\"]*)\" высота$")
+    public void setWindowSize(String widthRaw, String heightRaw) {
+        int width = Integer.valueOf(widthRaw);
+        int height = Integer.valueOf(heightRaw);
+        WebDriverRunner.getWebDriver().manage().window().setSize(new Dimension(width, height));
+        akitaScenario.write("Установлены размеры окна браузера: ширина " + widthRaw + " высота" + heightRaw);
+    }
+
+    /**
+     * Устанавливает размеры окна браузера
+     */
+    @И("^установлено разрешение экрана (\\d+) х (\\d+)$")
+    public void setBrowserWindowSize(int width, int height) {
+        getWebDriver().manage().window().setSize(new Dimension(width, height));
+        akitaScenario.write("Установлены размеры окна браузера: ширина " + width + " высота" + height);
+    }
+
+    /**
+     * Разворачивает окно с браузером на весь экран
+     */
+    @Если("^окно развернуто на весь экран$")
+    public void expandWindowToFullScreen() {
+        getWebDriver().manage().window().maximize();
+    }
+
+    /**
+     * Проверка выражения на истинность
+     * выражение из property, из переменной сценария или значение аргумента
+     * Например, string1.equals(string2)
+     * OR string.equals("string")
+     * Любое Java-выражение, возвращающие boolean
+     */
+    @Тогда("^верно, что \"([^\"]*)\"$")
+    public void expressionExpression(String expression) {
+        akitaScenario.getVars().evaluate("assert(" + expression + ")");
+    }
+
+    /**
+     * Переход на страницу по клику и проверка, что страница загружена
+     */
+    @И("^выполнен переход на страницу \"([^\"]*)\" после нажатия на (?:ссылку|кнопку) \"([^\"]*)\"$")
+    public void urlClickAndCheckRedirection(String pageName, String elementName) {
+        akitaScenario.getCurrentPage().getElement(elementName).click();
+        loadPage(pageName);
+        akitaScenario.write(" url = " + url());
+    }
+
+    /**
+     * Выполняется переход в конец страницы
+     */
+    @И("^совершен переход в конец страницы$")
+    public void scrollDown() {
+        Actions actions = new Actions(getWebDriver());
+        actions.keyDown(Keys.CONTROL).sendKeys(Keys.END).build().perform();
+        actions.keyUp(Keys.CONTROL).perform();
+    }
+
+    /**
+     * Проверка, что на странице не отображаются редактируемые элементы, такие как:
+     * -input
+     * -textarea
+     */
+    @Тогда("^открыта read-only форма$")
+    public void openReadOnlyForm() {
+        int inputsCount = getDisplayedElementsByCss("input").size();
+        assertTrue("Форма не read-only. Количество input-полей: " + inputsCount, inputsCount == 0);
+        int textareasCount = getDisplayedElementsByCss("textarea").size();
+        assertTrue("Форма не read-only. Количество элементов textarea: " + textareasCount, textareasCount == 0);
+    }
+
+    private List<SelenideElement> getDisplayedElementsByCss(String cssSelector) {
+        return $$(cssSelector).stream()
+            .filter(SelenideElement::isDisplayed)
+            .collect(Collectors.toList());
+    }
+
+    /**
+     * Выполняется поиск нужного файла в папке /Downloads
+     * Поиск осуществляется по содержанию ожидаемого текста в названии файла. Можно передавать регулярное выражение.
+     * После выполнения проверки файл удаляется
+     */
+    @Тогда("^файл \"(.*)\" загрузился в папку /Downloads$")
+    public void testFileDownloaded(String fileName) {
+        File downloads = getDownloadsDir();
+        File[] expectedFiles = downloads.listFiles((files, file) -> file.contains(fileName));
+        assertNotNull("Ошибка поиска файла", expectedFiles);
+        assertFalse("Файл не загрузился", expectedFiles.length == 0);
+        assertTrue(String.format("В папке присутствуют более одного файла с одинаковым названием, содержащим текст [%s]", fileName),
+            expectedFiles.length == 1);
+        deleteFiles(expectedFiles);
+    }
+
+    /**
+     * Скроллит экран до появления элемента. Полезно, если сайт длинный и элемент может быть не виден.
+     */
+    @Deprecated
+    @Тогда("^экран проскроллен до элемента \"([^\"]*)\"")
+    public void scrollToElement(String elementName) {
+        akitaScenario.getCurrentPage().getElement(elementName).scrollTo();
+    }
+
+    /**
+     * Скроллит экран до нужного элемента, имеющегося на странице, но видимого только в нижней/верхней части страницы.
+     */
+    @Тогда("^страница прокручена до элемента \"([^\"]*)\"")
+    public void scrollPageToElement(String elementName) {
+        akitaScenario.getCurrentPage().getElement(elementName).scrollTo();
+    }
+
+    /**
+     * Скроллит страницу вниз до появления элемента каждую секунду.
+     * Если достигнут футер страницы и элемент не найден - выбрасывается exception.
+     */
+    @И("^страница прокручена до появления элемента \"([^\"]*)\"$")
+    public void scrollWhileElemNotFoundOnPage(String elementName) {
+        SelenideElement el = null;
+        do {
+            el = akitaScenario.getCurrentPage().getElement(elementName);
+            if (el.exists()) {
+                break;
+            }
+            executeJavaScript("return window.scrollBy(0, 250);");
+            sleep(1000);
+        } while (!atBottom());
+        assertThat("Элемент " + elementName + " не найден", el.isDisplayed());
+    }
+
+    /**
+     * Скроллит страницу вниз до появления элемента с текстом каждую секунду.
+     * Если достигнут футер страницы и элемент не найден - выбрасывается exception.
+     */
+    @И("^страница прокручена до появления элемента с текстом \"([^\"]*)\"$")
+    public void scrollWhileElemWithTextNotFoundOnPage(String expectedValue) {
+        SelenideElement el = null;
+        do {
+            el = $(By.xpath(getTranslateNormalizeSpaceText(expectedValue)));
+            if (el.exists()) {
+                break;
+            }
+            executeJavaScript("return window.scrollBy(0, 250);");
+            sleep(1000);
+        } while (!atBottom());
+        assertThat("Элемент с текстом " + expectedValue + " не найден", el.isDisplayed());
+    }
+
+    /**
+     * Выполняется запуск js-скрипта с указанием в js.executeScript его логики
+     * Скрипт можно передать как аргумент метода или значение из application.properties
+     */
+    @Когда("^выполнен js-скрипт \"([^\"]*)\"")
+    public void executeJsScript(String scriptName) {
+        String content = loadValueFromFileOrPropertyOrDefault(scriptName);
+        Selenide.executeJavaScript(content);
+    }
+
+    /*
+     * Проверка совпадения значения из переменной и значения и property
+     */
+    @Тогда("^значения из переменной \"([^\"]*)\" и из property файла \"([^\"]*)\" совпадают$")
+    public void checkIfValueFromVariableEqualPropertyVariable(String envVarible, String propertyVariable) {
+        assertThat("Переменные " + envVarible + " и " + propertyVariable + " не совпадают",
+            (String) akitaScenario.getVar(envVarible), equalToIgnoringCase(loadProperty(propertyVariable)));
+    }
+
+
+    /**
+     * Возвращает каталог "Downloads" в домашней директории
+     *
+     * @return
+     */
+    private File getDownloadsDir() {
+        String homeDir = System.getProperty("user.home");
+        return new File(homeDir + "/Downloads");
+    }
+
+    /**
+     * Удаляет файлы, переданные в метод
+     *
+     * @param filesToDelete массив файлов
+     */
+    private void deleteFiles(File[] filesToDelete) {
+        for (File file : filesToDelete) {
+            file.delete();
+        }
+    }
+
+    /**
+     * Возвращает нормализованный(без учета регистра) текст
+     */
+    private String getTranslateNormalizeSpaceText(String expectedText) {
+        StringBuilder text = new StringBuilder();
+        text.append("//*[contains(translate(normalize-space(text()), ");
+        text.append("'АБВГДЕЁЖЗИЙКЛМНОПРСТУФХЦЧШЩЪЫЬЭЮЯ', 'абвгдеёжзийклмнопрстуфхчшщъыьэюя'), '");
+        text.append(expectedText);
+        text.append("') or contains(translate(normalize-space(text()), ");
+        text.append("'ABCDEFGHIJKLMNOPQRSTUVWXYZ', 'abcdefghijklmnopqrstuvwxyz'), '");
+        text.append(expectedText);
+        text.append("')]");
+        return text.toString();
+    }
+
+
+    //TODO: вынести в общий класс (AkitaScenario?)
+
+    public String getPropertyOrStringVariableOrValue(String propertyNameOrVariableNameOrValue) {
+        String propertyValue = tryLoadProperty(propertyNameOrVariableNameOrValue);
+        String variableValue = (String) akitaScenario.tryGetVar(propertyNameOrVariableNameOrValue);
+
+        boolean propertyCheck = checkResult(propertyValue, "Переменная " + propertyNameOrVariableNameOrValue + " из property файла");
+        boolean variableCheck = checkResult(variableValue, "Переменная сценария " + propertyNameOrVariableNameOrValue);
+
+        return propertyCheck ? propertyValue : (variableCheck ? variableValue : propertyNameOrVariableNameOrValue);
+    }
+
+    private boolean checkResult(String result, String message) {
+        if (isNull(result)) {
+            log.warn(message + " не найдена");
+            return false;
+        }
+        log.info(message + " = " + result);
+        akitaScenario.write(message + " = " + result);
+        return true;
+    }
+
+}
